+++ griptape/drivers/prompt/cohere_prompt_driver.py
@@ -11,7 +11,7 @@ if TYPE_CHECKING:
     from cohere import Client
 
 
-@define
+@define(kw_only=True)
 class CoherePromptDriver(BasePromptDriver):
     """
     Attributes:
@@ -20,15 +20,13 @@ class CoherePromptDriver(BasePromptDriver):
         client: Custom `cohere.Client`.
     """
 
-    api_key: str = field(kw_only=True, metadata={"serializable": False})
-    model: str = field(kw_only=True, metadata={"serializable": True})
+    api_key: str = field(metadata={"serializable": False})
+    model: str = field(metadata={"serializable": True})
     client: Client = field(
-        default=Factory(lambda self: import_optional_dependency("cohere").Client(self.api_key), takes_self=True),
-        kw_only=True,
+        default=Factory(lambda self: import_optional_dependency("cohere").Client(self.api_key), takes_self=True)
     )
     tokenizer: BaseTokenizer = field(
-        default=Factory(lambda self: CohereTokenizer(model=self.model, client=self.client), takes_self=True),
-        kw_only=True,
+        default=Factory(lambda self: CohereTokenizer(model=self.model, client=self.client), takes_self=True)
     )
 
     def try_run(self, prompt_stack: PromptStack) -> TextArtifact:
@@ -54,12 +52,18 @@ class CoherePromptDriver(BasePromptDriver):
     def _base_params(self, prompt_stack: PromptStack) -> dict:
         user_message = prompt_stack.inputs[-1].content
 
-        history_messages = [self._prompt_stack_input_to_message(input) for input in prompt_stack.inputs[:-1]]
+        history_messages = [
+            self._prompt_stack_input_to_message(input) for input in prompt_stack.inputs[:-1] if input.content
+        ]
 
-        return {
+        params = {
             "message": user_message,
-            "chat_history": history_messages,
             "temperature": self.temperature,
             "stop_sequences": self.tokenizer.stop_sequences,
             "max_tokens": self.max_tokens,
         }
+
+        if history_messages:
+            params["chat_history"] = history_messages
+
+        return params