+++ griptape/drivers/vector/base_vector_store_driver.py
@@ -2,12 +2,13 @@ from __future__ import annotations
 from abc import ABC, abstractmethod
 from concurrent import futures
 from dataclasses import dataclass
-from attrs import define, field, Factory
+from typing import Any
 from typing import Optional
+from attrs import define, field, Factory
 from griptape import utils
-from griptape.mixins import SerializableMixin
-from griptape.artifacts import TextArtifact
+from griptape.artifacts import TextArtifact, BaseArtifact, ListArtifact
 from griptape.drivers import BaseEmbeddingDriver
+from griptape.mixins import SerializableMixin
 
 
 @define
@@ -15,19 +16,19 @@ class BaseVectorStoreDriver(SerializableMixin, ABC):
     DEFAULT_QUERY_COUNT = 5
 
     @dataclass
-    class QueryResult:
+    class Entry:
         id: str
-        vector: Optional[list[float]]
-        score: float
+        vector: Optional[list[float]] = None
+        score: Optional[float] = None
         meta: Optional[dict] = None
         namespace: Optional[str] = None
 
-    @dataclass
-    class Entry:
-        id: str
-        vector: list[float]
-        meta: Optional[dict] = None
-        namespace: Optional[str] = None
+        @staticmethod
+        def from_dict(data: dict[str, Any]) -> BaseVectorStoreDriver.Entry:
+            return BaseVectorStoreDriver.Entry(**data)
+
+        def to_artifact(self) -> BaseArtifact:
+            return BaseArtifact.from_json(self.meta["artifact"])  # pyright: ignore[reportOptionalSubscript]
 
     embedding_driver: BaseEmbeddingDriver = field(kw_only=True, metadata={"serializable": True})
     futures_executor: futures.Executor = field(default=Factory(lambda: futures.ThreadPoolExecutor()), kw_only=True)
@@ -35,31 +36,40 @@ class BaseVectorStoreDriver(SerializableMixin, ABC):
     def upsert_text_artifacts(
         self, artifacts: dict[str, list[TextArtifact]], meta: Optional[dict] = None, **kwargs
     ) -> None:
-        utils.execute_futures_dict(
-            {
-                namespace: self.futures_executor.submit(self.upsert_text_artifact, a, namespace, meta, **kwargs)
-                for namespace, artifact_list in artifacts.items()
-                for a in artifact_list
-            }
-        )
+        with self.futures_executor as executor:
+            utils.execute_futures_dict(
+                {
+                    namespace: executor.submit(self.upsert_text_artifact, a, namespace, meta, **kwargs)
+                    for namespace, artifact_list in artifacts.items()
+                    for a in artifact_list
+                }
+            )
 
     def upsert_text_artifact(
-        self, artifact: TextArtifact, namespace: Optional[str] = None, meta: Optional[dict] = None, **kwargs
+        self,
+        artifact: TextArtifact,
+        namespace: Optional[str] = None,
+        meta: Optional[dict] = None,
+        vector_id: Optional[str] = None,
+        **kwargs,
     ) -> str:
-        if not meta:
-            meta = {}
+        meta = {} if meta is None else meta
+        vector_id = utils.str_to_hash(artifact.value) if vector_id is None else vector_id
 
-        meta["artifact"] = artifact.to_json()
-
-        if artifact.embedding:
-            vector = artifact.embedding
+        if self.does_entry_exist(vector_id, namespace):
+            return vector_id
         else:
-            vector = artifact.generate_embedding(self.embedding_driver)
+            meta["artifact"] = artifact.to_json()
 
-        if isinstance(vector, list):
-            return self.upsert_vector(vector, vector_id=artifact.id, namespace=namespace, meta=meta, **kwargs)
-        else:
-            raise ValueError("Vector must be an instance of 'list'.")
+            if artifact.embedding:
+                vector = artifact.embedding
+            else:
+                vector = artifact.generate_embedding(self.embedding_driver)
+
+            if isinstance(vector, list):
+                return self.upsert_vector(vector, vector_id=vector_id, namespace=namespace, meta=meta, **kwargs)
+            else:
+                raise ValueError("Vector must be an instance of 'list'.")
 
     def upsert_text(
         self,
@@ -69,13 +79,30 @@ class BaseVectorStoreDriver(SerializableMixin, ABC):
         meta: Optional[dict] = None,
         **kwargs,
     ) -> str:
-        return self.upsert_vector(
-            self.embedding_driver.embed_string(string),
-            vector_id=vector_id,
-            namespace=namespace,
-            meta=meta if meta else {},
-            **kwargs,
-        )
+        vector_id = utils.str_to_hash(string) if vector_id is None else vector_id
+
+        if self.does_entry_exist(vector_id, namespace):
+            return vector_id
+        else:
+            return self.upsert_vector(
+                self.embedding_driver.embed_string(string),
+                vector_id=vector_id,
+                namespace=namespace,
+                meta=meta if meta else {},
+                **kwargs,
+            )
+
+    def does_entry_exist(self, vector_id: str, namespace: Optional[str] = None) -> bool:
+        try:
+            return self.load_entry(vector_id, namespace) is not None
+        except Exception:
+            return False
+
+    def load_artifacts(self, namespace: Optional[str] = None) -> ListArtifact:
+        result = self.load_entries(namespace)
+        artifacts = [r.to_artifact() for r in result]
+
+        return ListArtifact([a for a in artifacts if isinstance(a, TextArtifact)])
 
     @abstractmethod
     def delete_vector(self, vector_id: str) -> None: ...
@@ -104,4 +131,4 @@ class BaseVectorStoreDriver(SerializableMixin, ABC):
         namespace: Optional[str] = None,
         include_vectors: bool = False,
         **kwargs,
-    ) -> list[QueryResult]: ...
+    ) -> list[Entry]: ...