+++ griptape/tools/vector_store_client/tool.py
@@ -1,32 +1,36 @@
-from typing import Optional
-from griptape.engines import VectorQueryEngine
+from __future__ import annotations
+from typing import Callable, Any
+from attrs import define, field, Factory
 from schema import Schema, Literal
-from attrs import define, field
-from griptape.artifacts import BaseArtifact, ErrorArtifact
+from griptape.artifacts import ErrorArtifact, BaseArtifact
+from griptape.artifacts import ListArtifact
+from griptape.drivers import BaseVectorStoreDriver
 from griptape.tools import BaseTool
 from griptape.utils.decorators import activity
 
 
-@define
+@define(kw_only=True)
 class VectorStoreClient(BaseTool):
     """
     Attributes:
         description: LLM-friendly vector DB description.
-        namespace: Vector storage namespace.
-        query_engine: `BaseQueryEngine`.
-        top_n: Max number of results returned for the query engine query.
+        vector_store_driver: `BaseVectorStoreDriver`.
+        query_params: Optional dictionary of vector store driver query parameters.
+        process_query_output_fn: Optional lambda for processing vector store driver query output `Entry`s.
     """
 
     DEFAULT_TOP_N = 5
 
-    description: str = field(kw_only=True)
-    query_engine: VectorQueryEngine = field(kw_only=True)
-    top_n: int = field(default=DEFAULT_TOP_N, kw_only=True)
-    namespace: Optional[str] = field(default=None, kw_only=True)
+    description: str = field()
+    vector_store_driver: BaseVectorStoreDriver = field()
+    query_params: dict[str, Any] = field(factory=dict)
+    process_query_output_fn: Callable[[list[BaseVectorStoreDriver.Entry]], BaseArtifact] = field(
+        default=Factory(lambda: lambda es: ListArtifact([e.to_artifact() for e in es]))
+    )
 
     @activity(
         config={
-            "description": "Can be used to search a vector database with the following description: {{ _self.description }}",
+            "description": "Can be used to search a database with the following description: {{ _self.description }}",
             "schema": Schema(
                 {
                     Literal(
@@ -40,6 +44,6 @@ class VectorStoreClient(BaseTool):
         query = params["values"]["query"]
 
         try:
-            return self.query_engine.query(query, top_n=self.top_n, namespace=self.namespace)
+            return self.process_query_output_fn(self.vector_store_driver.query(query, **self.query_params))
         except Exception as e:
             return ErrorArtifact(f"error querying vector store: {e}")