+++ tests/unit/tokenizers/test_simple_tokenizer.py
@@ -5,7 +5,7 @@ from griptape.tokenizers import SimpleTokenizer
 class TestSimpleTokenizer:
     @pytest.fixture
     def tokenizer(self):
-        return SimpleTokenizer(model="any model", max_input_tokens=1024, max_output_tokens=4096, characters_per_token=6)
+        return SimpleTokenizer(max_input_tokens=1024, max_output_tokens=4096, characters_per_token=6)
 
     def test_token_count(self, tokenizer):
         assert tokenizer.count_tokens("foo bar huzzah") == 3